requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
openai>=1.52.0
pandas==2.3.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv==1.0.0


